ocean-of-code: add buggy Map.cells_within_range()
It returns a lot of repeated cells. I think the BFS is not properly
checking visited cells.

ETA: cause the data structures deviated from the initial design of the
algorithm, and I'm not accounting for that.

I may want to come back to this version and make it work the future,
when I know a little more Rust.