Implement Add for Coord + dir
For the Rule of Cool.

Add moves the operands, doesn't take references. So I had to derive
Copy for Coord (again) and now every use of Coord + dir clones coord
instead of borrowing it.

To mitigate this, I changed Coord::after_move from a borrow method to a
move method, after_moving self in place.