Outline optimizations
I improved the logic in Grid::candidates_for to return early if it
becomes empty. Still not enough to pass the tests with large grids.

There's still lots of room for improvement though, my traversals are
inefficient, and I could narrow candidates down with lower bounds too.

Aside from a pesky bug that rears its head in large tests and/or tests
that include CrossSums cells.