Fix All The Bugs
Not really but I wanted to say that.

The ugliest bug is that I didn't quite understand how a Kankuro works.
The sum cells reset accumulation, meaning any sum cell to the left/above
a sum cell doesn't include the digits from the later... :facepalm:

I'll fix that next, in the meantime, this does solve quite a few bugs.
That I remember:

- sum_constrain! mutates candidates, so it needs select!, not select.
- it can't compute the min/max possible sum of other variables with
narrowed down candidates, so it needs the unused digits for the line.
This little fact motivated an overhaul (another) of constraint logic.