Fix buggy logic
Passes all tests :rocket:

It ain't pretty, but it does the job.

So I mistook the requirements. In my discharge, the problem statement
wasn't clear about this aspect, I learned about this in Wikipedia.

It turns out that, in Kakuros, sum cells in a direction reset the
constraints of uniqueness and sums in that direction. As if they were
defining entirely disjoint segments of the line they're in

Some lessons learned:

- Read The Fucking Man- Statement (plus Wikipedia)
- reminder that byebug is hella powerful