Add timeout-ing solution for Cooperative Mate
It passes the first 3 tests cases (shortest checkmate <= 3 moves) but
times out when the recursion tree explodes on the 4th (6 moves). It
takes 19s (3825 configurations tested) in my local. I haven't even tried
the 5th and 6th cases with 9 and 12 moves each.

There's some room for improvement but not in the order of magnitude that
I'd need to pass all tests, I think. In my defense, there aren't any
Ruby solutions for this puzzle as of yet.

What bothers me the most is the puzzle is supposed to take a dumb BFS,
without pruning or heuristics:
https://forum.codingame.com/t/community-puzzle-cooperative-mate-with-rook/189883/12

But then again, nobody asked me to write an expansive backtracking in
Ruby...

There are some silver linings though:

- I improved vscode's support for Ruby by a lot
- I iterated on the with_input method from kakuro_solver to allow
multiple test cases for one puzzle. Now I don't need to edit my-test.in,
run it, edit back if I want to perform regression tests, etc.