Byte Pair Encoding: massive reorganization
Move all the struct-centric method implementations to the end, leaving
just the headers at the top of the file.

Also, this is the first version that passes tests, cause I did away with
the stub returned at base case.

It actually passes all public tests except the last one, because of a
couple mystery characters at the end of the output.