Byte Pair Encoding: fix prc_to_s()
So I had to ensure that the string representation of the rules cons had
a null termination. It seems the absence of that null character at the
end is not an issue when printf-ing short enough strings...

Anyways, this passes all tests :rocket:. And I got the Multilinguist
achievement for it :tada:.

This is probably a botched job (all those memory leaks :see_no_evil:),
but then again, C is hard.